requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
pandas>=1.3.0
insightface>=0.7.3
onnxruntime>=1.8.0
redis>=4.3.0
python-dotenv>=0.19.0